Refactor and import protect counterfactual and intergrated gradient methods for tensorflow
