+++ alibi/explainers/anchors/anchor_image.py
