+++ alibi/explainers/anchors/anchor_tabular.py
@@ -1,5 +1,4 @@
 import copy
-# import logging
 from collections import OrderedDict, defaultdict
 from itertools import accumulate
 from typing import (Any, Callable, DefaultDict, Dict, List, Optional, Set,
