+++ alibi/saving.py
@@ -8,7 +8,7 @@ import warnings
 
 import dill
 import numpy as np
-import tensorflow as tf
+# import tensorflow as tf
 
 if TYPE_CHECKING:
     from alibi.api.interfaces import Explainer
@@ -117,7 +117,7 @@ def _simple_load(path: Union[str, os.PathLike], predictor, meta) -> 'Explainer':
     return explainer
 
 
-def _load_IntegratedGradients(path: Union[str, os.PathLike], predictor: Union[tf.keras.Model],
+def _load_IntegratedGradients(path: Union[str, os.PathLike], predictor: 'Union[tf.keras.Model]',
                               meta: dict) -> 'IntegratedGradients':
     layer_num = meta['params']['layer']
     if layer_num == 0:
